Move address parsing code to common module
