+++ loader/test/utils.test.js
@@ -9,7 +9,6 @@ const {
   filterObject,
   unpadNumber,
   getUniqueExternalIds,
-  parseUsAddress,
   RateLimit,
   parseUsPhoneNumber,
   cleanUrl,
@@ -237,30 +236,6 @@ describe("RateLimit", () => {
     expect(callTimes[1] - callTimes[0]).toBeLessThan(1050);
     expect(callTimes[2] - callTimes[1]).toBeGreaterThanOrEqual(995);
     expect(callTimes[2] - callTimes[1]).toBeLessThan(1050);
-  });
-});
-
-describe("parseUsAddress", () => {
-  it("Parses simple addresses", () => {
-    const parsed = parseUsAddress("3524 Somewhere St., Nowhere, NV 90210");
-    expect(parsed).toEqual({
-      lines: ["3524 Somewhere St."],
-      city: "Nowhere",
-      state: "NV",
-      zip: "90210",
-    });
-  });
-
-  it("Throws on invalid addresses", () => {
-    expect(() => {
-      parseUsAddress("Not an address or address-like string");
-    }).toThrow(ParseError);
-  });
-
-  it("Throws on invalid addresses that are structured like addresses", () => {
-    expect(() => {
-      parseUsAddress("., ., TX, 75244");
-    }).toThrow(ParseError);
   });
 });
